insertdata_vicroad/requirements.txt
Flask==2.0.2
elasticsearch==8.0.1
